requirements.txt
python-dotenv
fastapi
uvicorn
gunicorn
pydub